+++ internal/storage/postgresql/migrations/201_add_changelog_v5_4_2.down.sql
@@ -1,0 +1,1 @@
+DELETE FROM changelogs WHERE version = '5.4.2';